Fixes #5.
Running Dozer as wrapper in Flask.

Symptoms: When viewing http://localhost/_dozer/index, clicking TRACE link
on each of the objects results in a stacktrace similar to this:

-----
Traceback (most recent call last):
  File "/Users/mizhi/Development/python/virtualenvs/flask/lib/python2.7/site-packages/dozer/leak.py", line 97, in __call__
    return self.dowse(req)(environ, start_response)
  File "/Users/mizhi/Development/python/virtualenvs/flask/lib/python2.7/site-packages/dozer/leak.py", line 108, in dowse
    return method(req)
  File "/Users/mizhi/Development/python/virtualenvs/flask/lib/python2.7/site-packages/dozer/leak.py", line 220, in trace
    objid=str(objid or ''))
  File "/Users/mizhi/Development/python/virtualenvs/flask/lib/python2.7/site-packages/webob/response.py", line 373, in _body__set
    raise TypeError(msg)
TypeError: You cannot set Response.body to a text object (use Response.text)
-----

The problem is that the body property on the Response class is defined so i
that the setter rejects anything that is not a bytes object by raising the
TypeError.

Fix: Instead of Response.body, set Response.text.