Move the local virtualenv into .venv/
Reason: virtualenv creates a lot of clutter at the root (things like
pip-selftest.log etc.)

Keep a tree of symlinks in bin/, since I'm used to bin/foo working.
Note: do this instead of symlinking the entire bin/ because bin/python
breaks when bin/ is symlinked but there's no lib/ next to it.