Drop use of deprecated pkg_resources
In theory the replacement is importlib.resources.  In practice that
means dealing with the possibility of needing to extract the data files
into a temporary directory and then clean up afterwards, which is not
going to work for multiple reasons (other bits of the code already
inspect __file__ and expect things like Mako templates to live
alongside), so let's just keep doing what we were doing before
pkg_resources.

Maybe modernize things a bit by using pathlib.